requirements.txt
charset-normalizer==3.4.4
pillow==12.0.0
rectpack==0.2.2
reportlab==4.4.4
